src/requirements.txt
livekit-agents
livekit-plugins-cartesia
livekit-plugins-deepgram
livekit-plugins-openai
livekit-plugins-turn-detector
livekit-plugins-silero
livekit-plugins-google
livekit-plugins-noise-cancellation
mem0ai
duckduckgo-search
langchain_community
requests
python-dotenv